+++ dataforge-data/src/commonMain/kotlin/space/kscience/dataforge/data/GroupRule.kt
@@ -52,7 +52,9 @@ public interface GroupRule {
                 scope.launch {
                     set.updates.collect { name ->
                         val data = set.getData(name)
-                        val tagValue = data?.meta[key].string ?: defaultTagValue
+
+                        @Suppress("NULLABLE_EXTENSION_OPERATOR_WITH_SAFE_CALL_RECEIVER")
+                        val tagValue = data?.meta[key]?.string ?: defaultTagValue
                         map.getOrPut(tagValue) { ActiveDataTree(set.dataType) }.emit(name, data)
                     }
                 }